+++ agents/src/audio.ts
@@ -1,9 +1,26 @@
 // SPDX-FileCopyrightText: 2024 LiveKit, Inc.
 //
 // SPDX-License-Identifier: Apache-2.0
+import ffmpegInstaller from '@ffmpeg-installer/ffmpeg';
 import { AudioFrame } from '@livekit/rtc-node';
+import ffmpeg from 'fluent-ffmpeg';
+import type { ReadableStream } from 'node:stream/web';
 import { log } from './log.js';
+import { createStreamChannel } from './stream/stream_channel.js';
 import type { AudioBuffer } from './utils.js';
+
+ffmpeg.setFfmpegPath(ffmpegInstaller.path);
+
+export interface AudioDecodeOptions {
+  sampleRate?: number;
+  numChannels?: number;
+  /**
+   * Audio format hint (e.g., 'mp3', 'ogg', 'wav', 'opus')
+   * If not provided, FFmpeg will auto-detect
+   */
+  format?: string;
+  abortSignal?: AbortSignal;
+}
 
 export function calculateAudioDurationSeconds(frame: AudioBuffer) {
   // TODO(AJS-102): use frame.durationMs once available in rtc-node
@@ -71,4 +88,118 @@ export class AudioByteStream {
     this.#buf = new Int8Array(); // Clear buffer after flushing
     return frames;
   }
+}
+
+/**
+ * Decode an audio file into AudioFrame instances
+ *
+ * @param filePath - Path to the audio file
+ * @param options - Decoding options
+ * @returns AsyncGenerator that yields AudioFrame objects
+ *
+ * @example
+ * ```typescript
+ * for await (const frame of audioFramesFromFile('audio.ogg', { sampleRate: 48000 })) {
+ *   console.log('Frame:', frame.samplesPerChannel, 'samples');
+ * }
+ * ```
+ */
+export function audioFramesFromFile(
+  filePath: string,
+  options: AudioDecodeOptions = {},
+): ReadableStream<AudioFrame> {
+  const sampleRate = options.sampleRate ?? 48000;
+  const numChannels = options.numChannels ?? 1;
+
+  const audioStream = new AudioByteStream(sampleRate, numChannels);
+  const channel = createStreamChannel<AudioFrame>();
+  const logger = log();
+
+  // TODO (Brian): decode WAV using a custom decoder instead of FFmpeg
+  const command = ffmpeg(filePath)
+    .inputOptions([
+      '-probesize',
+      '32',
+      '-analyzeduration',
+      '0',
+      '-fflags',
+      '+nobuffer+flush_packets',
+      '-flags',
+      'low_delay',
+    ])
+    .format('s16le') // signed 16-bit little-endian PCM to be consistent cross-platform
+    .audioChannels(numChannels)
+    .audioFrequency(sampleRate);
+
+  let commandRunning = true;
+
+  const onClose = () => {
+    logger.debug('Audio file playback aborted');
+
+    channel.close();
+    if (commandRunning) {
+      commandRunning = false;
+      command.kill('SIGKILL');
+    }
+  };
+
+  const outputStream = command.pipe();
+  options.abortSignal?.addEventListener('abort', onClose, { once: true });
+
+  outputStream.on('data', (chunk: Buffer) => {
+    const arrayBuffer = chunk.buffer.slice(
+      chunk.byteOffset,
+      chunk.byteOffset + chunk.byteLength,
+    ) as ArrayBuffer;
+
+    const frames = audioStream.write(arrayBuffer);
+    for (const frame of frames) {
+      channel.write(frame);
+    }
+  });
+
+  outputStream.on('end', () => {
+    const frames = audioStream.flush();
+    for (const frame of frames) {
+      channel.write(frame);
+    }
+    commandRunning = false;
+    channel.close();
+  });
+
+  outputStream.on('error', (err: Error) => {
+    logger.error(err);
+    commandRunning = false;
+    onClose();
+  });
+
+  return channel.stream();
+}
+
+/**
+ * Loop audio frames from a file indefinitely
+ *
+ * @param filePath - Path to the audio file
+ * @param options - Decoding options
+ * @returns AsyncGenerator that yields AudioFrame objects in an infinite loop
+ */
+export async function* loopAudioFramesFromFile(
+  filePath: string,
+  options: AudioDecodeOptions = {},
+): AsyncGenerator<AudioFrame, void, unknown> {
+  const frames: AudioFrame[] = [];
+  const logger = log();
+
+  for await (const frame of audioFramesFromFile(filePath, options)) {
+    frames.push(frame);
+    yield frame;
+  }
+
+  while (!options.abortSignal?.aborted) {
+    for (const frame of frames) {
+      yield frame;
+    }
+  }
+
+  logger.debug('Audio file playback loop finished');
 }
