--- agents/src/codecs/index.ts
@@ -1,6 +1,0 @@
-// SPDX-FileCopyrightText: 2025 LiveKit, Inc.
-//
-// SPDX-License-Identifier: Apache-2.0
-import { audioFramesFromFile, loopAudioFramesFromFile } from './utils.js';
-
-export { audioFramesFromFile, loopAudioFramesFromFile };
